Moved CassandraContainerSetup to top level
